+++ packages/app-lixi-api/src/modules/account/account-cache.service.ts
@@ -1,4 +1,4 @@
-import { Account, AccountDana, AccountType } from '@bcpros/lixi-models';
+import { Account, AccountDana, AccountType, Role } from '@bcpros/lixi-models';
 import { InjectRedis } from '@songkeys/nestjs-redis';
 import { decode, encode } from '@msgpack/msgpack';
 import { Injectable, Logger } from '@nestjs/common';
@@ -54,7 +54,8 @@ export class AccountCacheService {
         avatar: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountAvatarImageUploadable?.uploads[0]),
         cover: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountCoverImageUploadable?.uploads[0]),
         hash160: dbValue.hash160.toString('hex'),
-        accountType: dbValue.accountType as AccountType
+        accountType: dbValue.accountType as AccountType,
+        role: dbValue.role as Role
       });
       await this.redis.hset(this.keyPrefix, id, Buffer.from(encode(account)));
       return account;
@@ -110,7 +111,8 @@ export class AccountCacheService {
             avatar: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountAvatarImageUploadable?.uploads[0]),
             cover: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountCoverImageUploadable?.uploads[0]),
             hash160: dbValue.hash160.toString('hex'),
-            accountType: dbValue.accountType as AccountType
+            accountType: dbValue.accountType as AccountType,
+            role: dbValue.role as Role
           });
           accountsMap.set(dbValue.id.toString(), account);
           const buffer = encode(account);
@@ -160,7 +162,8 @@ export class AccountCacheService {
         avatar: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountAvatarImageUploadable?.uploads[0]),
         cover: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountCoverImageUploadable?.uploads[0]),
         hash160: dbValue.hash160.toString('hex'),
-        accountType: dbValue.accountType as AccountType
+        accountType: dbValue.accountType as AccountType,
+        role: dbValue.role as Role
       });
 
       await this.redis.hset(this.keyPrefix, address, Buffer.from(encode(account)));
@@ -217,7 +220,8 @@ export class AccountCacheService {
             avatar: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.avatar?.upload),
             cover: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.cover?.upload),
             hash160: dbValue.hash160.toString('hex'),
-            accountType: dbValue.accountType as AccountType
+            accountType: dbValue.accountType as AccountType,
+            role: dbValue.role as Role
           });
           accountsMap.set(dbValue.address, account);
           const buffer = encode(account);
@@ -266,7 +270,8 @@ export class AccountCacheService {
         avatar: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountAvatarImageUploadable?.uploads[0]),
         cover: toImageUrl(this.deliveryUrl, this.cfAccountHash, dbValue.accountCoverImageUploadable?.uploads[0]),
         hash160: dbValue.hash160.toString('hex'),
-        accountType: dbValue.accountType as AccountType
+        accountType: dbValue.accountType as AccountType,
+        role: dbValue.role as Role
       });
 
       await this.redis.hset(this.keyPrefix, mnemonicHash, Buffer.from(encode(account)));