+++ packages/redux-store/src/store/escrow/escrow-order/escrow-order.generated.ts
@@ -118,6 +118,7 @@ export type GetModeratorAccountQuery = {
     rankScore?: number | null;
     telegramId?: string | null;
     telegramUsername?: string | null;
+    role: Types.Role;
     pages?: Array<{ __typename?: 'Page'; id: string; name: string }> | null;
     accountDana?: {
       __typename?: 'AccountDana';
@@ -163,6 +164,7 @@ export type GetRandomArbitratorAccountQuery = {
     rankScore?: number | null;
     telegramId?: string | null;
     telegramUsername?: string | null;
+    role: Types.Role;
     pages?: Array<{ __typename?: 'Page'; id: string; name: string }> | null;
     accountDana?: {
       __typename?: 'AccountDana';