+++ packages/lixi-models/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bcpros/lixi-models
 
+## 1.2.24
+
+### Patch Changes
+
+- Add role
+
 ## 1.2.23
 
 ## 1.2.22